334269: Contains several changes to Object Mapping Editor viewers.
 * Removes usage of DecoratingCellLabelProvider from Object Mapping
Editor: This was causing Viewer data to be "updated" with stale values
*after* being correctly updated.
 * Clears selection of all TreeViewers when saving the Object Mapping
Editor. This ensures that stale objects can not "wander" into the
Viewer via the old selection.
 * Uses the Viewer's Compararer (if available) in the Content
Provider's model listener. This should help to prevent issues
that arise when the Viewer considers the model objects equal even
though, technically, they are not.
 * Adds "delete-orphan" semantic to the Object Mapping Category /
Object Mapping Association relationship. This was a setting that
caused problems when used with Hibernate, but does not seem to cause
problems when used with EclipseLink.