Performance and responsiveness improvements when opening Projects.
Changed the order in which Project data is preloaded. The nodes now
come before the containers. The containers were implicity loading the
nodes as needed, rather than allowing all nodes to be loaded at once.

Modified Project Name caching to allow for caching of "unavailable"
(null) names. This greatly increases performance when loading large
Projects that reuse one or more other Projects that do not exist in
the database.