fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=336695
The key used to remove Test Data was not of the correct type. The
actual key for the Test Data map is of type String, but the key for
removal was of type Locale. This has been resolved by calling
toString() on the Locale object (as is also done elsewhere in the
changed method).