fixes 339060
A JPQL query was failing due to a missing whitespace. This changeset
adds the necessary whitespace. Also, it introduces a convention for
the location of whitespaces in a string that spans multiple lines of
source code. In such a concatenated string, necessary whitespaces are
always included at the beginning of a line of code. This makes it easy
to spot missing whitespace, as the text alignment will look "wrong" in
that case.