Various enhancements to monitoring support.
Adds "Source Directory" and "Installation Directory" fields to the AUT
Configuration dialog. Monitoring support uses these in order to
locate class and source files.

Adds "reset" functionality to monitoring agents. This allows
monitoring values to be reset before a Test Suite runs.

Monitoring reports that are larger than a defined threshold are no
longer sent to the to the client. Instead, overly large reports are
saved by the AUT Agent into the tmp directory. The path to the saved
file is persisted with the Test Result Summary's monitoring values.