Remote Control: Removes mouse movement after window activation.
After performing the click to activate a window, Swing, SWT, and RCP
RCs were moving the mouse pointer to the center of the activated
window. This can be misleading to someone observing the test, as it
appears as though the click was performed in the center of the
window. I have removed this additional mouse movement, as it does not
seem to serve any purpose other than confusion.