Changes context of all Jubula keybindings to "window".
Many of the bindings had a "dialogAndWindow" context.
They were changed in order to prevent, for example, attempts to open
multiple Project Properties dialogs.

The changes led to a binding conflict between "Toggle Pause on Error"
and "Show from Clipboard", which was resolved by changing the binding
for "Show from Clipboard".