Removes automatic activation of Properties View.
Removes automatic activation of the Properties View when opening one
or more Test Case Editors via double click in the Test Case Browser.
This functionality was causing some odd behavior when trying to open a
Test Case Editor while the Properties View is minimized. It was also
an annoyance when the user really had no interest in seeing the
Properties View just because they were opening a Test Case Editor.