fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=343846
Adds a new keybinding scheme "Functional Testing Default", which is
the default scheme for Jubula and GUIdancer. The bindings that were
causing the error messages (F3, M1+M2+T, and M1+M2+A) have been
assigned to the new keybinding scheme. This means that Jubula as a
plug-in has no default keybindings for "Revert Changes", "Open Test
Case", and "Edit Parameters".