fix for https://bugs.eclipse.org/bugs/show_bug.cgi?id=335665
For SWT/RCP: 
Replaced several calls (all calls in the implclasses package) to
Display.syncExec(...) with calls to
IEventThreadQueuer.invokeAndWait(...), as the latter method performs
additional error handling.

For SWT/RCP and Swing:
Changed the listeners used for synchronizing on conditions to handle
runtime exceptions outside of the EDT. This means that the default
exception handler for the EDT is no longer invoked, thus preventing
errors from bubbling all the way up to the UI. Instead, an exception
that occurs is rethrown when evaluating the synchronization condition,
allowing Jubula's RC to handle the exception.