Corrects equality check for RC navigation of Swing JTrees.
The previous implementation used "==", which caused nodes that
were technically equal but not exactly the same (e.g. Integer or
String objects) to not be recognized as equal. This caused infinite
loops when the RC tried to navigate a JTree that contained such
objects. The new implementation uses ObjectUtils.equals(...), which
ensures that such object will be recognized as equal and will prevent
the infinite loops that would otherwise occur. 
This changeset has essentially no effect for nodes that do not
override equals(), like DefaultMutableTreeNode.