fix for https://bugs.eclipse.org/bugs/show_bug.cgi?id=346069
Various operations (deletion, changes to Parameters, etc.) on Event
Handlers require modifications to the Test Case containing the Event
Handler, but this "parent" field was not always getting properly set
in the Event Handler object (for example, when opening a Test Case
Editor). I have resolved the primary symptom of the problem, in that
the "parent" field is now properly set when the Event Handler is
accessed by a content provider.