Fixes dependencies for rc.rcp.
The superfluous entries (which were only used during the build
process to manage dependencies automatically) in the "Require-Bundle"
section of the manifest were causing classloading issues when starting
an RCP AUT that actually contained the listed plugins (for example,
Jubula).

To fix this, the dependencies have been removed from the
"Require-Bundle" section of the manifest. This makes automated
dependency management no longer possible for rc.rcp. Also, the order
in which modules are built becomes important, hence the change to the
order of included plugins in the feature.xml.