Implements TabItem bounds calculation for SWT RC on Mac OS X Cocoa.
We were returning null for anything other than win32 and GTK, but SWT
has offered getBounds() as API for TabItem since 3.4. We now use
reflection to try to use the getBounds() method. Only if that fails do
we resort to the win32- and GTK-only implementations. This means that
win32 and GTK should work as far back as SWT 3.1, but all other
platforms are only supported for SWT 3.4 and higher.