Adds workaround for TabItem's getBounds() behavior on Mac OS X Cocoa.
The y-value provided by TabItem's getBounds() was far greater (lower
on the screen) than the actual bounds. This resulted in clicking
somewhere far below the TabItem. In the test scenario that I used, the
TabItem's actual y-value was 86, but getBounds() was reporting a
y-value of 351. 

I (successfully) tested the workaround on Mac OS X Cocoa, Windows
win32, and Linux GTK.