Moves Keyboard Mappings from server to client.
The Mappings are now transmitted from the client to the server upon
connection. This is the first step toward a solution for
https://bugs.eclipse.org/bugs/show_bug.cgi?id=346414, as it will allow
the reduction of the list of available Keyboard Layouts from a
complete list of Locales (whether they are available or not; does not
including variants such as MAC) to a list containing only those
Keyboard Layouts that are available.