Removes unavailable Keyboard Layouts from AUT Config dialog.
The list of Keyboard Layouts is no longer constructed by inserting all
available Locales. It is now representative of the actual available
Keyboard Mappings. Each Keyboard Mapping is still displayed as a
Locale, if it can be parsed as such. Otherwise, it is represented by
the Keyboard Mapping's file name.