fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=351565
Changes the way in which the rc.rcp bundle is activated. Previously,
we attempted to "stop()" the bundle during activation. Upon closer
inspection of the JavaDoc, this method should not be called by
clients.

Now, instead of trying to stop the bundle during activation,
we allow the bundle to run regardless of whether the host application
is running in an AUT context. However, the various listeners mentioned
in the bug report are only added if the host application is running in
an AUT context.