fixes https://bxapps.bredex.de/bugzilla/show_bug.cgi?id=129
The problem seems to have been some kind of deadlock on MacOS X when
waiting for the Display to process its pending requests. Since an OS
check was already implemented (checking for Windows), I have added an
additional check for MacOS. I have also added a check that determines
whether the "shake mouse" gesture is even necessary in order to trigger
a drag operation.