Adds Parameter details to Properties View and exported Test Results.
The Parameter details are included in exported Test Result XML files,
but not the corresponding HTML files.

I also noticed that the method of loading Parameter details referenced
the corresponding Test Step. I have changed this so that the Parameter
data is copied at execution time. This should prevent unexpected
behavior if, for example, something about the Test Step changes (or if
the Test Step itself is deleted).