Fixes an error that prevents editing AUT Configurations.
The error message stated that the plug-in for the AUT Toolkit could not
be loaded, but the actual source of the problem was that the storage of
the default AUT Agent preferences was recently changed to an
incompatible format. The default port is now retrieved directly via the
Java constant, rather than via the preferences.