fixes https://bxapps.bredex.de/bugzilla/show_bug.cgi?id=155
Under Windows, the cascade was being closed (via the Escape key) one
menu at a time, with an "is-menu-visible" check after each key
press/release. The problem was that the first opened menu was recognized
as *not* visible even though it *was* visible. As a result, Jubula
ended the Test Step as successful, even though the AUT's menu was not
successfully closed.
This seems to be a general problem with SWT 3.7. At least, the problem
did not occur in SWT 3.6.1.

The fix was to use a listener rather than a direct call to isVisible().