fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=344374
When programmatically selecting an item from a context menu, the menu is
now hidden *before* the selection event is fired for the item. This
guarantees that the context menu is no longer visible when the selection
event is processed. This does not seem to have a negative impact on
processing the selection event.