Fixes https://bxapps.bredex.de/bugzilla/show_bug.cgi?id=216.
Under specific circumstances in SWT/RCP AUTs, components were not being
found after an event that revealed a component with a parent that was
already visible. A good example of this is a CTabFolder that switches
its contents.

This changeset adds a null-check that prevents orphaning the revealed
component immediately after adding it.