Fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=370718.
Due to the way in which the model string for FunctionTokens is
concatenated, Reference arguments with no corresponding Parameter in the
parent Test Case were being evaluated as "null". The parameter converter
was then caching that result, which then resulted in "null" being
persisted in the Test Data instead of the Parameter GUID.
The instantiation of a new parameter converter prevents the caching in
this case, allowing for a re-evaluation of the model string for each
such Reference argument.