Fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=373914.
Whenever the last Language in a Project would be removed, an info dialog
is shown instead, informing the user that they cannot remove the last
Language in a Project. I chose this solution instead of disabling the
"Remove all" button because this solution also works for the "Remove"
button (as well as double-click / space bar) and also because the pick
list / slush bucket widget used is not very configurable with respect to
enablement.