Fixes https://bxapps.bredex.de/bugzilla/show_bug.cgi?id=201.
Addresses the fact that "command + a" (used to select the entire
contents of a text field / combo box) sometimes results in "a"
being entered into the text field on Mac. The workaround is to
immediately fall back to programmatically selecting the
contents of the text field / combo box on Mac.