partial fix for https://bugs.eclipse.org/bugs/show_bug.cgi?id=335660
Removes the caching of Test Results in the master session.
Test Results are now loaded into a separate session (per Viewer),
which is then closed when the Viewer is disposed.
The caching behavior is controlled by a preference, but the default
behavior is to disable caching. The behavior is determined by the
value of the preference at the time that the Viewer is opened.