fixes https://bugs.eclipse.org/bugs/show_bug.cgi?id=375655
The log listener registered during activation of the o.e.j.client.ui.rcp bundle
seems to have been the cause of the problem. As it uses UI elements to report
errors, it should never be registered when running a purely headless
application. Unfortunately, it looks as though this bundle may be accidentally
activated in some circumstances (which I was unable to reproduce) while running
a headless Jubula application (ex. dbtool, testexec).

To resolve the issue, I have moved the log listener registration code from
the bundle activator to Jubula application startup. A side-effect of this is
that general exceptions during job execution will no longer be handled by
Jubula's custom exception handler (i.e. no Jubula error dialog) when Jubula
is running as a Feature / Plug-in (ex. Jubula installed into IDE, Eclipse for
Testers). I think that this actually makes the exception handling a bit more
clear, so I don't see it as a drawback.