fixes https://bxapps.bredex.de/bugzilla/show_bug.cgi?id=337
Adds a feature project to the toolkit extension examples. This feature
project allows the client-side extension project to be exported as an
update site. This in turn allows proper p2-based installation of the
extension.

Also updates the Extension Manual to reflect the addition of the new
feature project and the new method of exporting and installing the
extension.