Replaces o.e.help.ui dependency with o.e.help dependency.
This should resolve the build failures that Jubula was causing
in Kepler simrel. The version number for o.e.help.ui was
increased due to API-breaking changes in Lucene.

The dependency was transitive (o.e.help.ui transitively
re-exports o.e.help), and it turns out that o.e.help.ui is
itself no longer required (if it ever was).

Various comments regarding the issue are available at:
http://dev.eclipse.org/mhonarc/lists/cross-project-issues-dev/msg08036.html