Generalizes error handling for unsupported Actions / Operations in RC.
The previous error handling would present a Configuration Error (with no
specific information) if the method corresponding to an Action could not
be found. The new error handling presents an Action Error with the 
message "Operation not supported by the selected toolkit." This
generalizes the implementation that is already used in toolkits with
unsupported concrete Actions (throwing a StepExecutionException with
the same message).