Removes return type of IEventThreadQueuer.invokeLater(...).
An asynchronously executed method should not have a return type
unless there is a framework in place to handle it (see
o.e.core.runtime.jobs.Job.ASYNC_FINISH for an example).
Otherwise, the returned Object is most likely to be null, as the
operation to invoke has not completed.

I've updated all references to this method in Jubula, without
problems because all references ignored the return value anyway.

It is worth noting that the Swing RC implementation executed such
operations synchronously if the current Thread was the EDT. This
is no longer the case; all invokeLater calls are now invoked
asynchronously on the EDT.

This commit also adds a default implementation for
IEventThreadQueuer.