Fixes http://bugs.eclipse.org/387995#c5 part A.
As I noticed that this behavior is also demonstrated by many
other dialogs in Jubula, I have also fixed these as well.

I fixed the problem by introducing the concept of an Editor
Operation, which is an operation that is executed within the
context of an editor, with setup and cleanup performed before
and after, respectively, as necessary.

While fixing this bug, I also noticed some variables and
methods that are no longer used (specifically, anything
having to do with INodePO.isReused()). I removed these
methods and variables, as they are no longer needed and
they were making code for certain situations (which
apparently also no longer occur) unnecessarily complex.