Fixes http://bugs.eclipse.org/389264.
The selection change was caused by the editor input being changed (via
reOpenEditor(...)) immediately after saving. Since the Selection Service
does not track the selection of non-active parts, the Properties View
was not being updated to match the new selection.

I've resolved this by no longer resetting the editor input for Test Case
/ Test Suite Editors immediately after saving. I am not sure why this
reset was there in the first place, which makes me somewhat leery of
removing it. However, I cannot at the moment see any reason why it would
be necessary, and there are at least two facts that support this
supposition:
 1. EditSupport#saveWorkVersion() (which is called during save) already
performs the operations that are necessary to ensure that the editor
contents are in a pristine state. For example, the Parameter Name and
Component Name Mappers are reset.
 2. No other Jubula editor uses reOpenEditor(...) in anything other than
an error case.