Fixes http://bugs.eclipse.org/391744.
Empty string test data was being handled in a special way
that was intended to be suitable for showing to the user.
Essentially, the empty string was represented as "''". Recent
changes caused this representation to also actually be used for
test execution, causing any tests relying on empty string test
data to fail.

I have corrected this for test execution, as well as for
displaying to the user. This commit causes Jubula to show an
actual empty string as representation for empty string test data.
This is a change in behavior that affects how Test Results are
displayed. It is not retroactive: Test Results generated by
Jubula before this changeset will continue to represent empty
string test data as "''".