http://eclip.se/430512 - JavaFX RC: Missing visibility checks
Adds a (recursive by parent) visibility check when checking whether a
component is "available". This means that non-visible components can be
present in the AUT hierarchy (and therefore may also be present when
determining the context of a neighboring component). Such non-visible
components will not, however, be found when waiting for attempting to
interact with said component.

I believe that this is similar to the behavior in other RCs.