RC: Adds a check to ignore source-bundles when building a classpath.
Since Jubula added source-bundles to the build, it has been impossible
to start an AUT from the embedded AUT Agent of an ITE started from the
IDE (well that was a mouthful). This bothered me enough to put in a fix
for it. Theoretically, this check could be expanded such that only valid
JAR files are included, but for now it seems prudent to take a
conservative approach:
If its a source-bundle JAR, don't include it on the classpath. If
*anything* goes wrong while trying to determine whether its a
source-bundle JAR, then just include it on the classpath.