Adds tracing for Component not Found errors.
The goal is to provide additional information to help track down
concurrency issues in the Swing RC. The tracing is currently only usable
in Swing RC and requires a System Property to be set to "true". The
System Property is not intended for production use, should not be
documented in the User Manual, and may be removed at any time.