Non-sprint task - adding recovering of missing test cases
Since we seem to lose somtimes references to existing test cases, this
should check all SpecTestCases from the project if they have a reference
on each loading of a project. If there are TestCases found they will be
added to the root node of the project.

see http://bugs.eclipse.org/455483 for more information

Change-Id: If8ad0b4cb505807e898d632f4a60ac4bb831c0db