Non-sprint task - fix for Bug 488218
We are now persisting the Unicode null symbol to the database instead of
empty strings. This is necessary due to the fact that in an Oracle
database an empty string is treated as null and therefore in the list
are gaps between the data which are persisted that way.

see http://eclip.se/488218 for more information

Change-Id: If72e0fcd63490c43ab76f56f852e133c6abb7984