Swing - Change all menu item actions to use doClick on MacOS
Since we had a lot of problems with the context menu and normal menu
tests we decided to change the test not do do clicks via the robot,
but doing the click programmatically. This works also for native mac
menus but they are not working completly as exspected. They are opend in
the top of the application windows instead unter the native menu.

Change-Id: I93380b9006d02a237806bc7dc7f838cd030cde7a