minor fixes of performance problems
the change in the ExecTestCasePO is speeding up loading of a test case
which as referenced central test data sets that have more parameters
than used by the test case. The method stated that there
are unused test data, but since we do not want to change ctds this is
ok. So we ignore central test data sets and therefore no save is done
at the opening of a test case if there are central test data sets.

The other change is just searching for references places in the project
which is loaded. Since otherwise it can be super slow to search in all
project for reference places to the CTDS which is also not possible.

Change-Id: Ided2fd362bafe78813a8d9cfc8bfb96ac4fb1564