Adding timeouts to most of the checks
To most of the checks timeouts are added, which have a polling delay of
100 ms and can be changed with the TEST_RC_POLLING_DELAY property.
Added a converter and the possibility to add more converter for the JSON
import.
Modified the API generation in the way there is also generated a method
which does not have the timeout parameter and automatically uses the
default 0.
HTML is not using any timout value at this moment

Change-Id: I36e9e22cc74b83266d386bd491e316a4e67ba08a