Add Workaround for JavaFx pulse not started
In some circumstances there might be the issue that the JavaFx pulse is 
not yet started but we already want to listen for changes to the 
Windows. This leads to a NullpointerException. Trying these multiple 
times after another did not work. It is now possible to have a static 
wait before our AUTserver initializes completely. You cannot set 
JUBULA_FX_WAIT_BEFORE_INIT environmental variable which takes an integer
as argument. This is the wait time before init in milliseconds

Change-Id: Ib1caa4a6d67da4b0d790b545ba8a79b09253d3ac